refactor(WebAPI): Replace IUpdateManager with UpdateManager in endpoints
